main(): In the change-of-address section, we only want to show the "you are already using that email address" message if the newaddress matches the case-preserved (subscribed) address.
Also, in the set_address section, if cpuser is None, set it to the the
user address, since that's what we'll use now as the old address in
the ChangeMemberAddress() call.

This and related changes should fix problems when the address we're
changing to differs for the current address by case only.